Decompose tasks pane into tasks-parts; codify split rule
Split panes/tasks.tsx (1112 LOC) into a sibling tasks-parts/ dir:
data.ts (pure cadence/draft/elapsed helpers + types + constants, with
data.test.ts), badges.tsx, task-running.tsx, task-fill.tsx, and
task-editor.tsx. tasks.tsx (now 324 LOC) keeps the public TasksView and
its create/edit/fill/save/delete orchestration.

Migrate the embedded fill + editor popups off their hand-rolled
backdrop/header/Esc scaffold onto the shared <Modal> shell, dropping the
local createBackdropClose/keydown wiring. The editor's delete-confirm Esc
ordering is preserved by gating its onClose on the confirm being closed.

Add a stated component-split convention (>~400 LOC or multiple
sub-components -> <name>-parts/, fixed data.ts/icons.tsx/per-component
sub-file vocabulary) in AGENTS.md Conventions and as internals invariant
12, turning the ad-hoc -parts/ pattern into a documented rule.
