+++ AGENTS.md
@@ -97,6 +97,7 @@ macOS and Windows are unaffected.
 - Run `make format` after every code change.
 - All UI must work without per-OS CSS branches (Electron + Chromium = same renderer everywhere).
 - **CSS design system.** Every text `font-size` uses a `--text-*` px token (`--text-3xs` 10px … `--text-3xl` 24px) — no rem ladders, no half-pixel sizes, no raw px on text (icon glyphs are the only exemption and say so in a comment). There is **no** spacing scale: padding/gap/margin are raw px on purpose (the tokenised scales are type, radius, and the modal width tiers). The dark palette is one `--dark-*` source in `styles.css` that both theme paths remap onto — never re-type dark hex literals. Define tokens before using them (`--danger`, `--text-dim`, `--preview-*` are real); never fake one with `var(--made-up, #literal)`. Shared shape primitives (`.section-header`, `.pill`) live in `src/renderer/primitives.css`. Full spec: [internals invariant 11](docs/explanation/internals.md#css-design-system).
+- **Component split threshold.** A renderer pane/modal `.tsx` past **~400 LOC** or holding more than one sub-component is split into a sibling `<name>-parts/` directory; `<name>.tsx` keeps the public component + orchestration and imports the rest. Fixed sub-file vocabulary: `data.ts` (pure logic + types + constants, dependency-free, with a `data.test.ts`), `icons.tsx` (pane-local glyphs), and one presentational `.tsx` per sub-component. A `-parts/` modal renders through the shared `<Modal>` shell. The threshold is a smell, not a gate — a genuinely single-responsibility long file (e.g. `index-tree.ts`) stays whole. Full spec: [internals invariant 12](docs/explanation/internals.md#component-parts).
 - In-window drag must use pointer events (`pointerdown` / `pointermove` / `pointerup` + `setPointerCapture`), never HTML5 drag-and-drop — Chromium's Wayland Ozone backend (forced for crisp text on Wayland sessions) silently breaks HTML5 DnD. The Projects-pane status drag follows this; terminal-tab and settings-row reorder are still HTML5 DnD (broken on Wayland) until converted. See [internals invariant 8](docs/explanation/internals.md#8-in-window-drag-uses-pointer-events-not-html5-drag-and-drop).
 - File-path conventions: every path crossing the IPC boundary is normalised to forward-slash form via `toPosix(p)` from `src/shared/path.ts`. Internal `fs` calls keep the native separator (use `path.join`); the renderer can then split on `/` without per-OS branches. Cross-OS shell wrapping (`bash -lc` / `cmd.exe /d /s /c` / `pwsh -Command`) lives in `src/main/terminals.ts`'s `wrapForShell`.
 - Spawned-process PATH: every pty / child process condash spawns takes its PATH from `spawnEnv()` in `src/main/shell-env.ts`, which resolves the user's **login-shell PATH** once at boot (`$SHELL -lic`, cached, 5 s timeout) so GUI-launched condash (Wayland / macOS — no `~/.profile` sourcing) still finds user-installed CLIs (`opencode`, `~/bin` wrappers). PATH only; other vars keep inherited values so the env-hygiene scrub stays intact. Wired into `terminals.ts` (pty + `force_stop`) and `launchers.ts`; warmed in `index.ts` at `whenReady`.
